requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
groq
reportlab
plotly
google-api-python-client